+++ test/Spec.hs
@@ -14,8 +14,6 @@ import qualified Test.Langchain.Chain.RetrievalQASpec as RetrievalQATest
 import qualified Test.Langchain.DocumentLoader.Core as DocumentLoaderTest
 import qualified Test.Langchain.DocumentLoader.CsvSpec as CsvLoaderTest
 import qualified Test.Langchain.DocumentLoader.DirectoryLoader as DirectoryLoaderTest
-import qualified Test.Langchain.DocumentLoader.HtmlSpec as HtmlLoaderTest
-import qualified Test.Langchain.DocumentLoader.JsonSpec as JsonLoaderTest
 import qualified Test.Langchain.Embeddings.Core as EmbeddingsTest
 import qualified Test.Langchain.Error as ErrorTest
 import qualified Test.Langchain.Graph.CompilationSpec as GraphCompilationTest
@@ -95,8 +93,6 @@ main =
           , DocumentLoaderTest.tests
           , DirectoryLoaderTest.tests
           , CsvLoaderTest.tests
-          , JsonLoaderTest.tests
-          , HtmlLoaderTest.tests
           , MemoryTest.tests
           , SummaryMemoryTest.tests
           , EntityMemoryTest.tests